feat(io): add ItkWasmMultiscaleSpatialImage class
This class makes a MultiscaleSpatialImage with a single scale
with one big chunk.

+++ packages/io/src/ZarrMultiscaleSpatialImage.ts
@@ -147,7 +147,7 @@ const makeCoords = ({
         }
         coords.set(dim, coordsPerElement);
       }
-      return coords.get(dim) as Float32Array;
+      return coords.get(dim);
     },
     has(dim: Dimension) {
       return axes.includes(dim);
